requirements.txt
setuptools,
ez_tools,
pandas,
mlflow==2.2.2,
notebook,
numpy,
scikit-learn,
matplotlib,
#python-bioc
pyYaml,
tqdm,
ensure,
joblib,
types-pyYaml,
#box
mlflow,
#Flask
#Flask-Config
#-e .